Reorganize tests. Prework before adding fakerjs.spec.ts
