--- support/cart-page.ts
@@ -1,11 +1,0 @@
-import { expect, Locator, type Page } from '@playwright/test';
-
-export class ProductsPage {
-  readonly page: Page;
-
-  constructor(page: Page) {
-    this.page = page;
-  }
-
-
-}
