+++ tests/end-to-end.spec.ts
@@ -1,13 +1,15 @@
 import { test } from '@playwright/test';
 import { SharedObjects } from '../support/shared/shared-objects';
-import { Header } from '../support/shared/header';
-import { LoginPage } from '../support/shared/login-page';
-import { ProductsPage } from '../support/products-page';
+import { Header } from '../support/saucedemo/header';
+import { LoginPage } from '../support/saucedemo/login-page';
+import { CartPage } from '../support/saucedemo/cart-page';
+import { ProductsPage } from '../support/saucedemo/products-page';
 
 test.describe('saucedemo - end-to-end tests', async () => {
   let sharedObjects: SharedObjects,
       header: Header,
       loginPage: LoginPage,
+      cartPage: CartPage,
       productsPage: ProductsPage;
 
   test.beforeEach('login', async ({ page }) => {
@@ -17,6 +19,7 @@ test.describe('saucedemo - end-to-end tests', async () => {
     sharedObjects = new SharedObjects(page);
     header = new Header(page);
     loginPage = new LoginPage(page);
+    cartPage = new CartPage(page);
     productsPage = new ProductsPage(page);
 
     if (!USERID || !PASSWORD) {
@@ -41,7 +44,7 @@ test.describe('saucedemo - end-to-end tests', async () => {
 
     // go to cart and checkout
     await header.clickCartButton();
-    await sharedObjects.assertCartItem('Sauce Labs Bolt T-Shirt');
+    await cartPage.assertCartItem('Sauce Labs Bolt T-Shirt');
     await sharedObjects.clickButton('Checkout');
 
     // checkout: your information
@@ -51,7 +54,7 @@ test.describe('saucedemo - end-to-end tests', async () => {
     await sharedObjects.clickButton('Continue');
 
     // checkout: overview
-    await sharedObjects.assertCartItem('Sauce Labs Bolt T-Shirt');
+    await cartPage.assertCartItem('Sauce Labs Bolt T-Shirt');
     await sharedObjects.clickButton('Finish');
 
     // checkout complete
